Changed notifyDisconnected in router
The notifyDisconnected function in the routerObject has been changed to
now take the device itself as an object, and not the stringname. Also, all
of the (one) call to this protected function and the references to the
protected list it handles are changed.

This simplifies the code, eliminates the use for the device.getName()
command in most cases, and solves the == vs. equals() problem commonly
found with strings in Java.