Hard-hitting question commented in
A to-do list of sorts that needs evaluation. Most importantly is
concerning the disconnect() function of the read/write thread. If that can
be made to work and disconnect completely and properly, then the server&client threads can be overlooked again to make sure it just grabs whatever it can and connects it.