Remove "internal" messages
Need for internal messages removed, as well as the pertinent code
necessary for parsing them. Additional byte at the beginning of messages
delegated for this purpose not yet removed.

Also, this marks the assumption that the rwThreads set in the router
object has exactly one thread for each connected device. Further
scanning and troubleshooting should be done to assure this.